refactor(cli): promote `colors` and `logging` to top-level utilities
